[Encoding][NFC] Moving Encoding attr/enum to EncodingAttrs.*
The revision keeps `EncodingBase.td` simple. It follows the IREE core
dialect style, which creates `EncodingTypes.h` and implements the
details in `EncodingAttrs.cpp`.
