+++ compiler/src/iree/compiler/Dialect/Encoding/IR/EncodingDialect.h
@@ -9,6 +9,7 @@
 
 #include "mlir/IR/Dialect.h"
 #include "mlir/IR/OpDefinition.h"
+#include "mlir/Support/TypeID.h"
 
 // clang-format off: must be included after all LLVM/MLIR headers
 #include "iree/compiler/Dialect/Encoding/IR/EncodingDialect.h.inc" // IWYU pragma: keep
